+++ packages/components/src/components/input-password/input-password-utils.ts
@@ -1,5 +1,7 @@
+import type { FormFieldAriaAttribute } from '../../types';
 import type { FormState } from '../../utils/form/form-state';
 export type InputPasswordState = FormState;
+export type InputPasswordAriaAttribute = FormFieldAriaAttribute;
 
 export type InputPasswordChangeEventDetail = Event;
 export type InputPasswordBlurEventDetail = Event;