+++ packages/components/src/components/input-week/input-week-utils.ts
@@ -1,5 +1,7 @@
+import type { FormFieldAriaAttribute } from '../../types';
 import type { FormState } from '../../utils/form/form-state';
 export type InputWeekState = FormState;
+export type InputWeekAriaAttribute = FormFieldAriaAttribute;
 
 export type InputWeekChangeEventDetail = Event;
 export type InputWeekBlurEventDetail = Event;