+++ packages/components/src/components/multi-select/multi-select/multi-select-utils.ts
@@ -1,9 +1,11 @@
 import { forceUpdate } from '@stencil/core';
+import type { FormFieldAriaAttribute } from '../../../types';
 import type { SelectComponentsDropdownDirection } from '../../../utils';
 import type { FormState } from '../../../utils/form/form-state';
 import type { MultiSelectOptionInternalHTMLProps } from '../multi-select-option/multi-select-option-utils';
 
 export type MultiSelectState = FormState;
+export type MultiSelectAriaAttribute = FormFieldAriaAttribute;
 export type MultiSelectDropdownDirection = SelectComponentsDropdownDirection;
 export type MultiSelectOption = HTMLPMultiSelectOptionElement & MultiSelectOptionInternalHTMLProps;
 export type MultiSelectOptgroup = HTMLPOptgroupElement;