+++ packages/components/src/components/multi-select/multi-select/multi-select.tsx
@@ -14,12 +14,13 @@ import {
   State,
   Watch,
 } from '@stencil/core';
-import type { BreakpointCustomizable, PropTypes, ValidatorFunction } from '../../../types';
+import type { BreakpointCustomizable, PropTypes, SelectedAriaAttributes, ValidatorFunction } from '../../../types';
 import {
   AllowedTypes,
   attachComponentCss,
   debounce,
   FILTER_STATUS_ANNOUNCE_TIMEOUT,
+  FORM_FIELD_ARIA_ATTRIBUTES,
   FORM_STATES,
   getComboboxAriaAttributes,
   getFilterStatusMessage,
@@ -39,6 +40,7 @@ import {
   isUsableOption,
   type Option,
   optionListUpdatePosition,
+  parseAndGetAriaAttributes,
   SELECT_DROPDOWN_DIRECTIONS,
   setHighlightedSelectOption,
   throwIfElementIsNotOfKind,
@@ -54,6 +56,7 @@ import { messageId, StateMessage } from '../../common/state-message/state-messag
 import type { InputSearchInputEventDetail } from '../../input-search/input-search-utils';
 import { getComponentCss } from './multi-select-styles';
 import {
+  type MultiSelectAriaAttribute,
   type MultiSelectChangeEventDetail,
   type MultiSelectDropdownDirection,
   type MultiSelectOptgroup,
@@ -82,6 +85,7 @@ const propTypes: PropTypes<typeof MultiSelect> = {
   form: AllowedTypes.string,
   dropdownDirection: AllowedTypes.oneOf<MultiSelectDropdownDirection>(SELECT_DROPDOWN_DIRECTIONS),
   compact: AllowedTypes.boolean,
+  aria: AllowedTypes.aria<MultiSelectAriaAttribute>(FORM_FIELD_ARIA_ATTRIBUTES),
 };
 
 /**
@@ -154,6 +158,9 @@ export class MultiSelect {
 
   /** Associates the multi-select with a form element by its ID when not directly nested inside it. */
   @Prop({ reflect: true }) public form?: string; // The ElementInternals API automatically detects the form attribute
+
+  /** Sets additional ARIA attributes on the combobox to improve accessibility for screen readers. */
+  @Prop() public aria?: SelectedAriaAttributes<MultiSelectAriaAttribute>;
 
   /** Emitted when the multi-select loses focus. */
   @Event({ bubbles: false }) public blur: EventEmitter<void>;
@@ -347,6 +354,7 @@ export class MultiSelect {
           isDisabled={this.disabled}
         />
         <button
+          {...parseAndGetAriaAttributes(this.aria)}
           aria-invalid={this.state === 'error' ? 'true' : null}
           type="button"
           role="combobox"
@@ -355,7 +363,6 @@ export class MultiSelect {
           {...getComboboxAriaAttributes(
             this.isOpen,
             this.required,
-            hasLabel(this.host, this.label) && labelId,
             selectMessageId,
             selectDescriptionId,
             listboxId