+++ packages/components/src/utils/a11y/form-field-aria-attributes.spec.ts
@@ -1,0 +1,7 @@
+import { FORM_FIELD_ARIA_ATTRIBUTES } from './form-field-aria-attribute';
+
+describe('FORM_FIELD_ARIA_ATTRIBUTES', () => {
+  it('should match snapshot', () => {
+    expect(FORM_FIELD_ARIA_ATTRIBUTES).toMatchSnapshot();
+  });
+});