+++ packages/components/src/utils/a11y/select/select-aria.ts
@@ -4,13 +4,11 @@ import { setAriaIDREF } from '../a11y';
 export const getComboboxAriaAttributes = (
   isOpen: boolean,
   isRequired: boolean,
-  labelId: string,
   messageId: string,
   descriptionId: string,
   listboxId: string
 ): AriaAttributes => {
   return {
-    'aria-labelledby': labelId || null,
     'aria-describedby': setAriaIDREF(messageId, descriptionId),
     'aria-haspopup': 'listbox',
     'aria-expanded': isOpen ? 'true' : 'false',